requirements-dev.txt
# 開發環境專用依賴
# 
# 這個檔案包含只在開發環境中需要的套件
# 安裝方式：pip install -r requirements-dev.txt

# 基礎依賴
-r requirements.txt

# =============================================================================
# 開發和除錯工具
# =============================================================================

# Django 開發工具
django-debug-toolbar>=4.2.0       # 除錯工具欄
django-extensions>=3.2.0          # 管理命令擴展
django-silk>=5.0.0                # 性能分析

# 代碼品質和格式化
black>=23.0.0                     # 代碼格式化
flake8>=6.0.0                     # 代碼風格檢查
isort>=5.12.0                     # import 排序
mypy>=1.5.0                       # 類型檢查
pre-commit>=3.3.0                 # Git 預提交鉤子

# 測試工具
pytest>=7.4.0                     # 測試框架
pytest-django>=4.5.0              # Django 測試支援
pytest-cov>=4.1.0                 # 測試覆蓋率
pytest-mock>=3.11.0               # Mock 工具
factory-boy>=3.3.0                # 測試數據工廠

# 文檔生成
Sphinx>=7.1.0                     # 文檔生成
sphinx-rtd-theme>=1.3.0           # Read the Docs 主題

# 開發服務器增強
django-livereload-server>=0.4     # 自動重載

# API 測試
httpie>=3.2.0                     # 命令行 HTTP 客戶端

# 數據庫工具
django-admin-interface>=0.19.0    # 美化的管理界面
